Do not skip duplicate messages if retain: true
SmartThings occasionally loses sync with HA. If the MQTT bridge retains the history without verifying that the device changed its state, it will refuse to process new command messages from HA because it already believes that the SmartThings device is in that state. This change ignores the MQTT bridge internal state machine when `retain` is `false`.